Fjern ubrukte funksjoner og typer og rename behanldingstema til stønadstype
